requirements.txt
# Web framework
django>=4.0,<5.0

# Machine Learning / Deep Learning
tensorflow>=2.11.0,<3.0

# Audio processing
librosa>=0.10.0,<0.11.0
soundfile>=0.12.1

# Array and numerical computation
numpy>=1.23.0

# For file upload handling (optional, usually included with Django)
requests>=2.28.0
gunicorn
django-cors-headers
djangorestframework